pyproject.toml
[project]
name = "newsletter"
version = "0.1.0"
description = "package for automating an events newsletter"
readme = "README.md"
requires-python = ">=3.11"

packages = [{ include = "newsletter" }]

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.126.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
supabase = "^2.3.4"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
openai = "^1.78.0"
beautifulsoup4 = "^4.13.3"
pgeocode = "^0.5.0"
curl-cffi = "^0.10.0"
tenacity = "^9.1.2"
anthropic = "^0.51.0"
requests-html = "^0.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
